Add MoE training script, fix AMP dtype mismatch in expert routing
- scripts/train_moe.py: Full end-to-end MoE training with CTC + LID losses
  Supports --synth mode (on-the-fly rendering, no LMDB needed) and
  --train_dir mode (LMDB datasets). Tested on MPS, 2 scripts, loss decreasing.
- Fix RuntimeError in SWABlockMoE: expert MLP output dtype must match
  destination tensor under AMP (half→float cast)
- Same fix in ScriptCTCHeads forward

+++ src/model/moe_encoder.py
@@ -80,7 +80,7 @@ class ScriptCTCHeads(nn.Module):
                 continue
             head = self.heads[script]
             out, _ = head["lstm"](features[mask])
-            logits[mask] = head["proj"](out)
+            logits[mask] = head["proj"](out).to(logits.dtype)
 
         return logits
